fixed interrupt argument size

+++ src/api/interrupt.h
@@ -12,10 +12,10 @@ extern "C"
 {
 #endif
 
-extern uint8_t interrupt_enable(uint8_t interrupt);
-extern uint8_t interrupt_disable(uint8_t interrupt);
+extern uint8_t interrupt_enable(uint32_t interrupt);
+extern uint8_t interrupt_disable(uint32_t interrupt);
 extern uint8_t interrupt_register(uint32_t interrupt, void (*pfnHandler)(void));
-extern uint8_t interrupt_set_priority(uint8_t interrupt, uint8_t priority);
+extern uint8_t interrupt_set_priority(uint32_t interrupt, uint8_t priority);
 
 #ifdef __cplusplus
 }
